+++ src/main/java/com/example/demo/BorrowingsService.java
@@ -10,6 +10,10 @@ public class BorrowingsService {
     private List<BorrowedBook> borrowings;
     Customer customer;
     Book book;
+    private BookService bookService;
+    private CustomerService customerService;
+    private List<Customer> customers = customerService.getListOfCustomers();
+    private List<Book> books = bookService.getListOfBooks();
     public BorrowingsService(){
         this.borrowings = init();
     }
@@ -18,8 +22,8 @@ public class BorrowingsService {
     private List<BorrowedBook> init(){
         List<BorrowedBook> borrowings = new ArrayList<>();
         BorrowedBook borrowing1 = new BorrowedBook();
-        // customer = customers.get(0);
-        // book = books.get(0);
+        customer = customers.get(0);
+        book = books.get(0);
         borrowing1.setCustomer(customer);
         borrowing1.setBook(book);
         borrowings.add(borrowing1);
@@ -36,10 +40,8 @@ public class BorrowingsService {
     }
     //CREATING NEW BORROWING
     public String createBorrowing(Long customerId, Long bookId){
-//        Long bookId = borrowedBook.bookId;
-//        Long customerId = borrowedBook.customerId;
-//        customer = customers.get(Math.toIntExact(customerId));
-//        book = books.get(Math.toIntExact(bookId));
+        customer = customers.get(Math.toIntExact(customerId));
+        book = books.get(Math.toIntExact(bookId));
         BorrowedBook newBorrowing = new BorrowedBook();
         newBorrowing.setCustomer(customer);
         newBorrowing.setBook(book);
